Demo unit tests
Added basic module imports test and a check that the settings submod is
working. These demonstrate use of fixtures and a tests.data.<TEST> sub-
packaging approach I use.

Also, refined behavior of the linter to ignore star or unused imports as
we are not enforcing that degree of discipline at present.