Added jobs-history
This work uncovered a challenge with using three verbosity levels to gate
whether specific fields show up in a given formatted display. Essentially,
we have two display cases for history, which are both Lister-subclassed.
One is to show the table view, and the other is to recapitualte the JSON
view when executed with the -v flag. Added log levels and a per-class
attribute EXTRA_VERBOSITY which combine together to give finer-grained
control over what fields are returned.