Date searches for jobs plus general quality improvements for all search params
Jobs can now be searched by creation, end, or update date. Search by date string or
natural language (i.e. 'today') is supported. Date searches on apps and systems is not
working and it appears this is a service level issue.

All supported modifiers for a search param are now shown in the help string, with
the defaut being highlighted.

All values for a field only specific choices are available are now shown in the
help string. Validation for choices is currently broken though.

Defaults for apps, jobs, and systems have been curated and all combinations should work