Refactored dotenv loading into a submodule
In order to display the current discovered .env path, the value passed to
load_env needed to be captured. It also needed to be made clear in the code
what the search order was. The tapis_cli.settings.config#find_config searches
module path tree, $PWD, and $HOME for a .dot env file. The config#load_config
function loads the result of that path search. Having the load function
implemented as a callable function should make it easier to reload the
environment settings from within an interactive shell.

The net result of this work is that 'tapis settings list' now includes
an _ENV_PATH entry that shows where the CLI is looking for its settings.

Closes #209