TaccApiDirectClient now can report API server error messages
In order to capture and return the 'message' key from Tapis API
responses, I updated the TaccApiDirectClient class, which is what is
used to make calls in lieu of AgavePy. I wrapped the raise_for_status
in each of its HTTP method handlers in a function that extracts the
API server message and adds it to the HTTPError object raised when
a non-20x status code is encountered.

This behavior can be toggled by setting TAPIS_PY_VERBOSE_ERRORS to
a falsey value.