Use forked cookiecutter
Cookiecutter is a bit in transition now, and as a consequence, their last
release (1.7.0) lacks an advertised feature we built on, which is support
for sub-directories in a cookiecutter repo. To solve this, TACC forked
cookiecutter as tacc-cookiecutter-1.7.0.1 and released to PyPi. This is
obviously unsustainable but the cc project is dormant to fix requests
and we need to include it to support the CLI's templating capability"